Latent Action
Started with a latent action for moving because this can have multiple inputs. While it does have the drawback of not proceeding when it fails I do still think it might be a better approach. I will keep both versions, the Async Action and the Latent action and will make sure they are compatible later on
Duplicated the showcase level so I have one with many units and one with only one unit, this way I don't have to remove them